Adding Contact Form and Redoing Dropdown
While the title says that I did alot more. Make commits more often Reggie so we dont have to do this.

Anyways I added a Contact Form  using FormSubmit so that I wouldnt have to make a backend myself.  I still have to find some way to add a space inbetween  Input and Label, otherwise its in a good spot.

Dropdown was made to be toggle by the image I added, it jumps the  section instead of scrolling which is the only change that is really needed  for it.

Added a src/images to hold images that will be used in the proejct

I also added a text box around the words in each div in PortfolioLanding. These will be changed to fit the design later but it is in a good spot.

Global : Add responsivness  and shadows, otherwise were in a  good spot.